+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Change Log
 
+* 3.22.7
+    * Updated dependencies
+    * Allow for selecting a Default test level when deploying #373
 * 3.22.6
     * Fix retrieving from package with only one type. This resolves `Retrieve from package.xml feature doesn't retrieve settings metadata` #368
 * 3.22.5